refactor: ecomerce-service 디렉토리 구조 변경
- as-is : 계층형 / to-be : 도메인형
