+++ module-ecommerce-service/src/main/java/com/shop/ecommerceservice/product/dto/request/ProductCreateRequest.java
@@ -1,6 +1,6 @@
 package com.shop.ecommerceservice.product.dto.request;
 
-import com.shop.ecommerceservice.product.Product;
+import com.shop.ecommerceservice.product.entity.Product;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
