+++ module-ecommerce-service/src/main/java/com/shop/ecommerceservice/product/dto/response/ProductStockResponse.java
@@ -1,7 +1,7 @@
 package com.shop.ecommerceservice.product.dto.response;
 
-import com.shop.ecommerceservice.product.Product;
-import com.shop.ecommerceservice.product.ProductStock;
+import com.shop.ecommerceservice.product.entity.Product;
+import com.shop.ecommerceservice.product.entity.ProductStock;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
