+++ module-ecommerce-service/src/main/java/com/shop/ecommerceservice/product/repository/ProductStockRepository.java
@@ -1,5 +1,6 @@
-package com.shop.ecommerceservice.product;
+package com.shop.ecommerceservice.product.repository;
 
+import com.shop.ecommerceservice.product.entity.ProductStock;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
